refactor(core): extract maintenance/ai-structured.ts — AI structured-content backfill
